+++ eagle-jpm/eagle-jpm-web/src/main/webapp/app/apps/jpm/index.js
@@ -34,8 +34,8 @@
 		templateUrl: "partials/job/overview.html",
 		controller: "overviewCtrl",
 		resolve: { time: true }
-	}).route("jpmStatistic", {
-		url: "/jpm/statistic?startTime&endTime",
+	}).route("jpmStatistics", {
+		url: "/jpm/statistics",
 		site: true,
 		templateUrl: "partials/job/statistic.html",
 		controller: "statisticCtrl"
@@ -57,39 +57,37 @@
 		controller: "compareCtrl"
 	});
 
-	jpmApp.portal({name: "YARN Jobs", icon: "home", list: [
+	jpmApp.portal({name: "YARN Jobs", icon: "taxi", list: [
 		{name: "Overview", path: "jpm/overview"},
-		{name: "Job Statistic", path: "jpm/statistic"},
+		{name: "Job Statistics", path: "jpm/statistics"},
 		{name: "Job List", path: "jpm/list"}
 	]}, true);
 
 	jpmApp.service("JPM", function ($q, $http, Time, Site, Application) {
-		// TODO: mock auth
-		var _hash = btoa('eagle:secret');
-
-		var JPM = {};
+		var JPM = window._JPM = {};
 
 		// TODO: timestamp support
-		JPM.QUERY_LIST = '${baseURL}/eagle-service/rest/entities?query=${query}[${condition}]{${fields}}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}';
-		JPM.QUERY_GROUPS = '${baseURL}/eagle-service/rest/list?query=${query}[${condition}]<${groups}>{${field}}${order}${top}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}';
-		JPM.QUERY_GROUPS_INTERVAL = '${baseURL}/eagle-service/rest/list?query=${query}[${condition}]<${groups}>{${field}}${order}${top}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}&intervalmin=${intervalMin}&timeSeries=true';
-		JPM.QUERY_METRICS = '${baseURL}/eagle-service/rest/entities?query=GenericMetricService[${condition}]{*}&metricName=${metric}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}';
-		JPM.QUERY_METRICS_AGG = '${baseURL}/eagle-service/rest/list?query=GenericMetricService[${condition}]<${groups}>{${field}}${order}${top}&metricName=${metric}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}';
-		JPM.QUERY_METRICS_INTERVAL = '${baseURL}/eagle-service/rest/list?query=GenericMetricService[${condition}]<${groups}>{${field}}${order}${top}&metricName=${metric}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}&intervalmin=${intervalMin}&timeSeries=true';
-		JPM.QUERY_MR_JOBS = '${baseURL}/eagle-service/rest/mrJobs/search';
-		JPM.QUERY_JOB_LIST = '${baseURL}/eagle-service/rest/mrJobs?query=%s[${condition}]{${fields}}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}';
-		JPM.QUERY_TASK_STATISTIC = '${baseURL}/eagle-service/rest/mrJobs/${jobId}/taskCountsByDuration?site=${site}&timelineInSecs=${times}&top=${top}';
+		JPM.QUERY_LIST = '${baseURL}/rest/entities?query=${query}[${condition}]{${fields}}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}';
+		JPM.QUERY_GROUPS = '${baseURL}/rest/entities?query=${query}[${condition}]<${groups}>{${field}}${order}${top}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}';
+		JPM.QUERY_GROUPS_INTERVAL = '${baseURL}/rest/entities?query=${query}[${condition}]<${groups}>{${field}}${order}${top}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}&intervalmin=${intervalMin}&timeSeries=true';
+		JPM.QUERY_METRICS = '${baseURL}/rest/entities?query=GenericMetricService[${condition}]{*}&metricName=${metric}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}';
+		JPM.QUERY_METRICS_AGG = '${baseURL}/rest/entities?query=GenericMetricService[${condition}]<${groups}>{${field}}${order}${top}&metricName=${metric}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}';
+		JPM.QUERY_METRICS_INTERVAL = '${baseURL}/rest/entities?query=GenericMetricService[${condition}]<${groups}>{${field}}${order}${top}&metricName=${metric}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}&intervalmin=${intervalMin}&timeSeries=true';
+		JPM.QUERY_MR_JOBS = '${baseURL}/rest/mrJobs/search';
+		JPM.QUERY_JOB_LIST = '${baseURL}/rest/mrJobs?query=%s[${condition}]{${fields}}&pageSize=${limit}&startTime=${startTime}&endTime=${endTime}';
+		JPM.QUERY_JOB_STATISTIC = '${baseURL}/rest/mrJobs/jobCountsByDuration?site=${site}&timeDistInSecs=${times}&startTime=${startTime}&endTime=${endTime}';
+		JPM.QUERY_TASK_STATISTIC = '${baseURL}/rest/mrTasks/taskCountsByDuration?jobId=${jobId}&site=${site}&timeDistInSecs=${times}&top=${top}';
 
-		JPM.QUERY_MR_JOB_COUNT = '${baseURL}/eagle-service/rest/mrJobs/runningJobCounts';
-		//JPM.QUERY_MR_JOB_METRIC_TOP = '${baseURL}eagle-service/rest/mrJobs/jobMetrics/list';
+		JPM.QUERY_MR_JOB_COUNT = '${baseURL}/rest/mrJobs/runningJobCounts';
+		//JPM.QUERY_MR_JOB_METRIC_TOP = '${baseURL}eagle-service/rest/mrJobs/jobMetrics/entities';
 
 		/**
 		 * Fetch query content with current site application configuration
 		 * @param {string} queryName
 		 */
-		var getQuery = JPM.getQuery = function(queryName) {
+		var getQuery = JPM.getQuery = function(queryName, siteId) {
 			var baseURL;
-			var siteId = Site.current().siteId;
+			siteId = siteId || Site.current().siteId;
 			var app = Application.find("JPM_WEB_APP", siteId)[0];
 			var host = app.configuration["service.host"];
 			var port = app.configuration["service.port"];
@@ -136,16 +134,11 @@
 		}
 
 		JPM.get = function (url, params) {
-			$http.defaults.withCredentials = true;
-			var promise = $http({
+			return $http({
 				url: url,
 				method: "GET",
-				params: params,
-				headers: {'Authorization': "Basic " + _hash}
+				params: params
 			});
-
-			$http.defaults.withCredentials = false;
-			return promise;
 		};
 
 		JPM.condition = function (condition) {
@@ -358,6 +351,23 @@
 			return _list;
 		};
 
+		/**
+		 * Fetch job duration distribution
+		 * @param {string} site
+		 * @param {string} times
+		 * @param {{}} startTime
+		 * @param {{}} endTime
+		 */
+		JPM.jobDistribution = function (site, times, startTime, endTime) {
+			var url = common.template(getQuery("JOB_STATISTIC"), {
+				site: site,
+				times: times,
+				startTime: Time.format(startTime),
+				endTime: Time.format(endTime)
+			});
+			return JPM.get(url);
+		};
+
 		JPM.taskDistribution = function (site, jobId, times, top) {
 			var url = common.template(getQuery("TASK_STATISTIC"), {
 				site: site,
@@ -467,6 +477,7 @@
 	});
 
 	jpmApp.requireCSS("style/index.css");
+	jpmApp.require("widget/jobStatistic.js");
 	jpmApp.require("ctrl/overviewCtrl.js");
 	jpmApp.require("ctrl/statisticCtrl.js");
 	jpmApp.require("ctrl/listCtrl.js");
