+++ eagle-jpm/eagle-jpm-web/src/main/webapp/app/apps/jpm/partials/job/compare.html
@@ -18,7 +18,7 @@
 
 <div class="row flex">
 	<div class="col-sm-12 col-md-3">
-		<div class="box box-primary" style="height: 100%;">
+		<div class="box box-primary">
 			<div class="box-header with-border">
 				<h3 class="box-title">
 					Summary
@@ -53,7 +53,7 @@
 		</div>
 	</div>
 	<div class="col-sm-12 col-md-9">
-		<div class="box box-primary" style="height: 100%;">
+		<div class="box box-primary">
 			<div class="box-header with-border">
 				<h3 class="box-title">
 					Comparison
