+++ eagle-jpm/eagle-jpm-web/src/main/webapp/app/apps/jpm/partials/job/detail.html
@@ -18,7 +18,7 @@
 
 <div class="row flex">
 	<div class="col-lg-6 col-md-12">
-		<div class="box box-primary" style="height: 100%;">
+		<div class="box box-primary">
 			<div class="box-header with-border">
 				<h3 class="box-title">
 					Job Info
@@ -84,7 +84,7 @@
 	</div>
 
 	<div class="col-lg-6 col-md-12">
-		<div class="box box-primary" style="height: 100%;">
+		<div class="box box-primary">
 			<div class="box-header with-border">
 				<h3 class="box-title">
 					Map Reduce
