+++ eagle-jpm/eagle-jpm-web/src/main/webapp/app/apps/jpm/style/index.css
@@ -68,3 +68,9 @@
 	color: #000;
 	font-size: 30px;
 }
+
+.small-box.jpm {
+	margin: 0;
+	height: 100%;
+	min-height: 110px;
+}
