--- eagle-server/src/main/webapp/app/dev/apps/example/index.js
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-(function () {
-	/**
-	 * `register` is global function that for application to set up 'controller', 'service', 'directive', 'route' in Eagle
-	 */
-	var exampleApp = register(['ngRoute', 'ngAnimate', 'ui.router', 'eagle.service']);
-
-	exampleApp.route({
-		url: "/test",
-		templateUrl: "test.html",
-		controller: "testCtrl"
-	});
-
-	exampleApp.route("testState", {
-		url: "/test/123",
-		templateUrl: "test.html",
-		controller: "testCtrl"
-	});
-
-	exampleApp.controller('testCtrl', function ($scope) {
-		$scope.aaa = 123123123;
-	});
-})();
