--- eagle-server/src/main/webapp/app/dev/apps/jpm_dev/partials/job/task.html
@@ -1,149 +1,0 @@
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-  -->
-
-<div class="box box-primary">
-	<div class="box-header with-border">
-		<h3 class="box-title">
-			Task Schedule Trend
-		</h3>
-	</div>
-	<div class="box-body">
-		<div class="jpm-chart">
-			<div chart class="jpm-chart-container" series="scheduleSeries" category="scheduleCategory"></div>
-			<div ng-if="(scheduleSeries || []).length === 0" class="overlay">
-				<i class="fa fa-refresh fa-spin"></i>
-			</div>
-		</div>
-	</div>
-</div>
-
-<div class="nav-tabs-custom">
-	<ul class="nav nav-tabs">
-		<li class="active"><a href="#scheduleDistribution" data-toggle="tab">Schedule Distribution</a></li>
-		<li><a href="#durationDistribution" data-toggle="tab">Duration Distribution</a></li>
-	</ul>
-	<div class="tab-content keepContent">
-		<!-- By Schedule Distribution -->
-		<div class="tab-pane fade in active" id="scheduleDistribution">
-			<div class="row">
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="statusSeries" category="bucketScheduleCategory" option="statusOption"></div>
-						<div ng-if="(statusSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="durationSeries" category="bucketScheduleCategory" option="durationOption"></div>
-						<div ng-if="(durationSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="hdfsReadSeries" category="bucketScheduleCategory" option="hdfsReadOption"></div>
-						<div ng-if="(hdfsReadSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="hdfsWriteSeries" category="bucketScheduleCategory" option="hdfsWriteOption"></div>
-						<div ng-if="(hdfsWriteSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="localReadSeries" category="bucketScheduleCategory" option="localReadOption"></div>
-						<div ng-if="(localReadSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="localWriteSeries" category="bucketScheduleCategory" option="localWriteOption"></div>
-						<div ng-if="(localWriteSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-			</div>
-		</div>
-
-		<!-- By Duration Distribution -->
-		<div class="tab-pane fade" id="durationDistribution">
-			<div class="row">
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="durationStatusSeries" category="bucketDurationCategory" option="durationStatusOption"></div>
-						<div ng-if="(durationStatusSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="durationMapReduceSeries" category="bucketDurationCategory" option="durationMapReduceOption"></div>
-						<div ng-if="(durationMapReduceSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="durationHdfsReadSeries" category="bucketDurationCategory" option="durationHdfsReadOption"></div>
-						<div ng-if="(durationHdfsReadSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="durationHdfsWriteSeries" category="bucketDurationCategory" option="durationHdfsWriteOption"></div>
-						<div ng-if="(durationHdfsWriteSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="durationLocalReadSeries" category="bucketDurationCategory" option="durationLocalReadOption"></div>
-						<div ng-if="(durationLocalReadSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-				<div class="col-sm-12 col-md-6">
-					<div class="jpm-chart">
-						<div chart class="jpm-chart-container" series="durationLocalWriteSeries" category="bucketDurationCategory" option="durationLocalWriteOption"></div>
-						<div ng-if="(durationLocalWriteSeries || []).length === 0" class="overlay">
-							<i class="fa fa-refresh fa-spin"></i>
-						</div>
-					</div>
-				</div>
-			</div>
-		</div>
-	</div>
-</div>
