--- eagle-server/src/main/webapp/app/dev/apps/jpm_dev/style/index.css
@@ -1,76 +1,0 @@
-@CHARSET "UTF-8";
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-.jpm-chart {
-	position: relative;
-	margin-bottom: 15px;
-}
-
-.jpm-chart h3 {
-	margin: 10px 0 10px 0;
-}
-
-.jpm-chart .jpm-chart-container {
-	height: 300px;
-	position: relative;
-}
-
-.jpm-chart .jpm-chart-container.scroll {
-	overflow-y: auto;
-}
-
-.jpm-chart.chart-lg .jpm-chart-container {
-	height: 350px;
-}
-
-.with-border .jpm-chart {
-	padding-bottom: 15px;
-	margin-bottom: 15px;
-	border-bottom: 1px solid #f4f4f4;
-}
-
-.with-border .jpm-chart:last-child {
-	padding-bottom: 0;
-	margin-bottom: 0;
-	border-bottom: 0;
-}
-
-.jpm-chart .overlay {
-	top: 0;
-	bottom: 0;
-	position: absolute;
-	width: 100%;
-	background: rgba(255,255,255,0.7);
-}
-
-.jpm-chart .overlay > .fa {
-	position: absolute;
-	top: 50%;
-	left: 50%;
-	margin-left: -15px;
-	margin-top: -15px;
-	color: #000;
-	font-size: 30px;
-}
-
-.small-box.jpm {
-	margin: 0;
-	height: 100%;
-	min-height: 110px;
-}
